backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Docker & Jupyter
docker==6.1.3
jupyter-client==8.6.0
ipykernel==6.27.1
pyzmq==25.1.1

# LLM
openai==1.3.7
anthropic==0.7.7
tiktoken==0.5.1

# Git
GitPython==3.1.40

# Data Science
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0

# Utils
pydantic==2.5.2
pydantic-settings==2.1.0
aiofiles==23.2.1
